+++ gen_tree/requirements.txt
@@ -1,0 +1,2 @@
+scipy
+numpy
